refactor: "Bookshelf" and "Book" components added.
The new components encapsulate a lot of the JSX that was in the
top-level App. All books are passed to each bookshelf right now, and the
bookshelf filters the viewable books on its own. It might be more
logical to have them filtered at time of passing. I'm not sure yet that
it matters.

I have the beginnings of an onChange handler for the shelf changer
select list, but it's not yet expected to do anything useful. I'm just
experimenting trying to get the semantics straight right now. No
judgment.