feat: Book mover 90% functional, bookshelves now populated from Library state!
Selecting the desired bookshelf in the book mover select list now
appears to do everything right except it doesn't force refresh the page.
Manually reloading reflects the expected state.

In addition, I finally got the bookshelves in the data structure I
wanted initially, and found a way to populate it reliably in
componentDidLoad. As a bonus, it's the exact same structure returned by
BooksAPI.update(), so updating state after an update() is super simple.