refactor: Remove Library.state.bookshelves
I used a complicated object 'bookshelves', and it made everything
harder. Instead, I'm just using the 'books' returned from
BooksAPI.getAll() in my Library.state. Code is easier to debug & read
now.

I also introduce a couple of known bugs here:
 * bookshelf names are now the shortNames stored with the book instead
of human-friendly ones. I'll get to it tomorrow.
 * because I wanted my handling of bookshelf names to be generic, I have
a function that identifies unique shelf property values in the 'books'
array. But you can't do 'map()' on a Set, so I do Array.from on the set,
and Sets are not *ordered* collections, so when you move books around
now (which is totally functional), the bookshelves will change position.
Lame.