fix: Shelves were switching positions when books were moved
I wanted a more general approach to assembling the supported bookshelves
that would allow for easy expansion by the user later (a user could
easily add whatever shelves they wanted). That's not a requirement,
though, and assembling them on each render caused whole shelves to
switch position in the UI randomly, which is weird. I took that
functionality out and replaced it with a fixed mapping of short to
prettified shelf names in Bookshelf.js.

This also allowed for a bit of code cleanup in both Library and
Bookshelf.