fix: Query updates 'off-by-one' problems, and handling queries for unsupported terms
I was managing the query through this.state. Inside the onChange handler
for search input, I'd call this.setState, then immediately call the api
search, passing this.state.query, which hadn't been updated yet. The
result is that the search was always happening with the query resulting
from the *last* change -- it was always off by one. Taking 'query'
completely out of this.state simplifies the code in addition to solving
off-by-one issues.

I also added code to deal with unsupported api queries, which return a
generic js object rather than throwing an actual catchable exception, so
now when you search for an unsupported query, the result object is just
empty and there are no errors in the console.