fix: Handle missing cover images & authors attributes
On both the search page and the main page, if a book is missing a cover
image or authors attribute, it's handled by iterating over the results
and putting default values in for those attributes. I downloaded a 'no
cover image' icon to use for missing cover images.

As far as I can see right now, I have *one bug* left before I can
submit: the changeShelf dropdown does not change it's 'selected' value
when you change the shelf of a book in the search results. It's
infuriating.