fix: Search page bookmover select list not updating
When a user changed the shelf of a book from the search page, a rerender
would happen, after which the select list would not have the proper
shelf selected. That is fixed.

I also removed duplicate code in Bookshelf.js by simply using
conditional rendering to determine if a title is necessary instead of
duplicating the entire JSX output for each case (rendering the main
page, or rendering the search page).

I also refactored the changeShelf function, which is where my bug lived.
I separated local updates to books in this.state from API updates.